+++ test/Sentry.Unity.Tests/UnityLogExceptionTests.cs
@@ -7,6 +7,14 @@ namespace Sentry.Unity.Tests
 {
     public class UnityLogExceptionTests
     {
+        [Test]
+        public void ToSentryException_MarkedAsUnhandled()
+        {
+            var sentryException = new UnityLogException().ToSentryException();
+
+            Assert.IsFalse(sentryException.Mechanism?.Handled);
+        }
+
         [TestCaseSource(nameof(ParsingTestCases))]
         public void ToSentryException_ParsingTestCases(
             string logString,
@@ -103,6 +111,11 @@ UnityEngine.EventSystems.EventSystem:Update() (at /Applications/Unity/Hub/Editor
                                 InApp = true
                             },
                         }
+                    },
+                    Mechanism = new Mechanism
+                    {
+                        Handled = false,
+                        Type = "unity.log"
                     }
                 } },
             new object[] {
@@ -236,6 +249,11 @@ Expected: False == True",
                                 InApp = false
                             },
                         }
+                    },
+                    Mechanism = new Mechanism
+                    {
+                        Handled = false,
+                        Type = "unity.log"
                     }
                 } },
             // TODO: Current parsing logic isn't able to parse this (editor only stack trace)