+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Fixes
 
 - Android Native Support no longer crashes when built with Mono while trying to capture a crash ([#524](https://github.com/getsentry/sentry-unity/pull/524))
+- Bump Sentry CLI 1.72.0 ([#526](https://github.com/getsentry/sentry-unity/pull/526))
+  - [changelog](https://github.com/getsentry/sentry-cli/releases/tag/1.72.0)
+  - [diff](https://github.com/getsentry/sentry-cli/compare/1.71.0...1.72.0)
 - Bump Sentry .NET SDK 3.13.0 ([#503](https://github.com/getsentry/sentry-unity/pull/503))
   - [changelog](https://github.com/getsentry/sentry-dotnet/blob/3.13.0/CHANGELOG.md)
   - [diff](https://github.com/getsentry/sentry-dotnet/compare/3.12.3...3.13.0)